[editor canvas] Move RunController to CanvasController/Run.
